// helpers/types.ts
export enum eEthereumNetwork {
  buidlerevm = "buidlerevm",
  kovan = "kovan",
  ropsten = "ropsten",
  main = "main",
  coverage = "coverage",
  hardhat = "hardhat",
  tenderly = "tenderly",
  rinkeby = "rinkeby",
  goerli = "goerli",
  sepolia = "sepolia",
  arbitrum = "arbitrum"
}

export enum eMantleNetwork{
  mantleTestnet = "mantleTestnet",
  mantle = "mantle"
}